+++ esp/zicBlock/main/view.h
@@ -12,6 +12,6 @@ public:
     }
 
     virtual void render() = 0;
-    virtual void onEncoder(int id, int8_t direction, uint32_t tick) = 0;
+    virtual void onEncoder(int id, int8_t direction, uint64_t tick) = 0;
     virtual bool onKey(uint16_t id, int key, int8_t state) = 0;
 };
